Feature/add before hook implementation (#12)
* Update gemlock

This includes updates resulting from running bundle install on m1 macbook

* Add Before Hooks

This adds support for before hooks as defined in https://openfeature.dev/specification/sections/hooks, including refactoring the hook type by making it a sealed interface and adding fields according to spec, adding the Before Hook type, and adding behavior as specified by OpenFeature related to before hooks.

Adding calls to the before hooks within feature flag evaluation is intentionally left out of this commit

* Add spec for hook hints

As per spec, hook hints, which are basically a hash map, are a paramater to calls to hooks. This adds a spec to verify they are being passed through

* Add incorporation of before hooks into flag evaluation

Flag evaluation will invoke before hooks in the right place. The sequencing follows spec as well (API -> Client -> Invocation -> Provider).

Note that the hook hints are dropped for now. Also sequencing, while trivial, is not tested

* Add Requirement 4.1.2 from spec

'The hook context SHOULD provide: access to the client metadata and the provider metadata fields.'

This adds to the appopriate structs and updates test suites. It made sense to refactor the context construct within the tests into a helper builder function as well

* Refactor to break out hook sequencing and typing

This breaks out a type for all hooks into a separate struct type with assocaited helper methods for filter and sequencing. We also drop the other types in order to simplify for this stage, and apply T.cast to facilitate this simplification